refactor(core): clean up query builder code review findings
Remove dead _evaluating loading state, extract shared groupAnnotations
utility to eliminate duplication, and remove redundant type assertions
after isFilterGroup type guards.
