+++ packages/core/src/components/control-bar/annotation-select.ts
@@ -2,12 +2,7 @@ import { LitElement, html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { annotationSelectStyles } from './annotation-select.styles';
 import { handleDropdownEscape } from '../../utils/dropdown-helpers';
-import {
-  ANNOTATION_CATEGORIES,
-  TAXONOMY_ORDER,
-  type CategoryName,
-  type GroupedAnnotation,
-} from './annotation-categories';
+import { groupAnnotations, type GroupedAnnotation } from './annotation-categories';
 
 /**
  * Custom dropdown component for annotation selection with section headers and search
@@ -131,54 +126,10 @@ class ProtspaceAnnotationSelect extends LitElement {
   }
 
   /**
-   * Categorize annotations according to the plan
+   * Categorize annotations using the shared utility.
    */
   private categorizeAnnotations(annotations: string[]): GroupedAnnotation[] {
-    const categorized: Record<CategoryName, string[]> = {
-      UniProt: [],
-      InterPro: [],
-      Taxonomy: [],
-      Other: [],
-    };
-
-    for (const annotation of annotations) {
-      let found = false;
-      for (const [category, categoryAnnotations] of Object.entries(ANNOTATION_CATEGORIES)) {
-        if ((categoryAnnotations as readonly string[]).includes(annotation)) {
-          categorized[category as CategoryName].push(annotation);
-          found = true;
-          break;
-        }
-      }
-      if (!found) {
-        categorized.Other.push(annotation);
-      }
-    }
-
-    // Sort annotations within each category
-    categorized.UniProt.sort((a, b) => a.localeCompare(b));
-    categorized.InterPro.sort((a, b) => a.localeCompare(b));
-    categorized.Other.sort((a, b) => a.localeCompare(b));
-    // Taxonomy uses predefined order
-    categorized.Taxonomy.sort((a, b) => {
-      const aIndex = TAXONOMY_ORDER.indexOf(a as (typeof TAXONOMY_ORDER)[number]);
-      const bIndex = TAXONOMY_ORDER.indexOf(b as (typeof TAXONOMY_ORDER)[number]);
-      if (aIndex === -1 && bIndex === -1) return a.localeCompare(b);
-      if (aIndex === -1) return 1;
-      if (bIndex === -1) return -1;
-      return aIndex - bIndex;
-    });
-
-    // Build grouped array, sorting categories alphabetically (Other last)
-    const groups: GroupedAnnotation[] = [];
-    const categoryOrder: CategoryName[] = ['InterPro', 'Taxonomy', 'UniProt', 'Other'];
-    for (const category of categoryOrder) {
-      if (categorized[category].length > 0) {
-        groups.push({ category, annotations: categorized[category] });
-      }
-    }
-
-    return groups;
+    return groupAnnotations(annotations);
   }
 
   /**
